[Rename] 스타일 경로 수정
