+++ src/styles/home/NoteItem-styled.js
